[EasyEncryption] Resolves concerns.
