+++ packages/EasyEncryption/src/Bridge/Symfony/Resources/config/aws_pkcs11_encryptor.php
@@ -5,6 +5,8 @@ namespace Symfony\Component\DependencyInjection\Loader\Configurator;
 
 use EonX\EasyEncryption\AwsPkcs11Encryptor;
 use EonX\EasyEncryption\Bridge\BridgeConstantsInterface;
+use EonX\EasyEncryption\Builder\AwsCloudHsmSdkOptionsBuilder;
+use EonX\EasyEncryption\Configurator\AwsCloudHsmSdkConfigurator;
 use EonX\EasyEncryption\Interfaces\AwsPkcs11EncryptorInterface;
 
 return static function (ContainerConfigurator $containerConfigurator): void {
@@ -14,8 +16,7 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->autoconfigure();
 
     $services
-        ->set(AwsPkcs11EncryptorInterface::class, AwsPkcs11Encryptor::class)
-        ->arg('$userPin', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_USER_PIN))
+        ->set(AwsCloudHsmSdkOptionsBuilder::class)
         ->arg('$hsmCaCert', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_HSM_CA_CERT))
         ->arg(
             '$disableKeyAvailabilityCheck',
@@ -24,14 +25,21 @@ return static function (ContainerConfigurator $containerConfigurator): void {
         ->arg('$hsmIpAddress', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_HSM_IP_ADDRESS))
         ->arg('$cloudHsmClusterId', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_CLOUD_HSM_CLUSTER_ID))
         ->arg('$awsRegion', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_AWS_REGION))
-        ->arg('$awsRoleArn', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_AWS_ROLE_ARN))
-        ->arg('$aad', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_AAD))
         ->arg('$serverClientCertFile', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_SERVER_CLIENT_CERT_FILE))
         ->arg('$serverClientKeyFile', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_SERVER_CLIENT_KEY_FILE))
-        ->arg('$cloudHsmOptions', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_CLOUD_HSM_OPTIONS))
+        ->arg('$cloudHsmOptions', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_CLOUD_HSM_OPTIONS));
+
+    $services
+        ->set(AwsCloudHsmSdkConfigurator::class)
+        ->arg('$awsRoleArn', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_AWS_ROLE_ARN))
         ->arg(
             '$useCloudHsmConfigureTool',
             param(BridgeConstantsInterface::PARAM_AWS_PKCS11_USE_CLOUD_HSM_CONFIGURE_TOOL)
-        )
+        );
+
+    $services
+        ->set(AwsPkcs11EncryptorInterface::class, AwsPkcs11Encryptor::class)
+        ->arg('$userPin', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_USER_PIN))
+        ->arg('$aad', param(BridgeConstantsInterface::PARAM_AWS_PKCS11_AAD))
         ->arg('$defaultKeyName', param(BridgeConstantsInterface::PARAM_DEFAULT_KEY_NAME));
 };
